flush after newline (needed by qosd as parent process)

+++ include/Client.h
@@ -26,7 +26,7 @@ struct Args {
     uint8_t dscp_snd = 0;
     uint32_t num_samples = 10;
     uint32_t mean_inter_packet_delay = 200;
-    uint8_t timeout = 10;
+    uint32_t timeout = 10;
     uint32_t seed = 0;
     char sep = ',';
     bool sync_time = false;
